requirements.txt
# Transcription
openai-whisper>=20231117
torch>=2.0.0

# Vector database
chromadb>=0.4.22

# Embeddings and LLM
openai>=1.0.0
anthropic>=0.18.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.66.0
